http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@508579 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/stomp/StompSessionManager.cpp
@@ -67,7 +67,7 @@ StompSessionManager::~StompSessionManager(void)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-unsigned int StompSessionManager::getNextSessionId(void)
+long long StompSessionManager::getNextSessionId(void)
 {
     synchronized( &mutex )
     {
@@ -78,7 +78,7 @@ unsigned int StompSessionManager::getNextSessionId(void)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-unsigned int StompSessionManager::getNextConsumerId(void)
+long long StompSessionManager::getNextConsumerId(void)
 {
     synchronized( &mutex )
     {
